Refactor search
Add "contract" and move KiwixSearchWidget from utils
