+++ app/src/main/java/org/kiwix/kiwixmobile/KiwixMobileActivity.java
@@ -83,7 +83,7 @@ import org.kiwix.kiwixmobile.search.SearchActivity;
 import org.kiwix.kiwixmobile.settings.KiwixSettingsActivity;
 import org.kiwix.kiwixmobile.utils.DimenUtils;
 import org.kiwix.kiwixmobile.utils.DocumentParser;
-import org.kiwix.kiwixmobile.utils.KiwixSearchWidget;
+import org.kiwix.kiwixmobile.search.KiwixSearchWidget;
 import org.kiwix.kiwixmobile.utils.KiwixTextToSpeech;
 import org.kiwix.kiwixmobile.utils.LanguageUtils;
 import org.kiwix.kiwixmobile.utils.NetworkUtils;
